Moved intermediate taskflow to new directory.
Moved the new ClusterProvisioningTaskFlow intermediate taskflow class
to a new directory.
